refactor(common,engine): hoist allocatePort into @linkcode/common/node
