+++ packages/agent-adapter/package.json
@@ -15,6 +15,7 @@
     "@anthropic-ai/claude-agent-sdk": "^0.3.206",
     "@anthropic-ai/sdk": "^0.110.0",
     "@earendil-works/pi-coding-agent": "^0.80.6",
+    "@linkcode/common": "workspace:*",
     "@linkcode/schema": "workspace:*",
     "@linkcode/transport": "workspace:*",
     "@openai/codex": "^0.144.1",
